Added error handler functionality for GSL.
This allows for a user provided callback function from GSL on errors.
By default GSL calls `abort` when an error occurs.

I've added an type for GSL error handlers that is user overrideable. I've
also added a default standard GSL error handler that can be used to
create a stacktrace with `fatal` when an error occurs.

The `reset` of the error handler is currently not working.