Added a Makefile and an awk filter script for running the unit tests
This allows me to do something like this:

```
make test-gpolynomial
```

And have it run only the unit tests tagged with the string `gpolynomial` and to filter the
output to only show the interesting parts of the test. Otherwise, I get a bunch of "SKIPPED"
lines and empty lines that clutter the output.